requirements.txt
waitress~=2.1.2
Flask~=3.0.1
python-dotenv~=1.0.1
Dragon-Ball-API~=1.0.0